+++ activerecord/test/models/joke_comment.rb
@@ -1,0 +1,3 @@
+class JokeComment < ActiveRecord::Base
+  belongs_to :joke, class_name: 'AJoke', foreign_key: :joke_id
+end